1) moved router to separate file; 2) fixed spaces
