+++ src/components/ui/layout/header/index.tsx
@@ -1,9 +1,9 @@
 import { FC } from 'react'
 
 const Header: FC = () => {
-    return (
-        <div></div>
-    )
+  return (
+    <div></div>
+  )
 }
 
 export default Header;
