+++ src/main.tsx
@@ -1,26 +1,9 @@
 import ReactDOM from 'react-dom/client'
-import MainPage from './pages/main'
-import './index.css'
 import { store } from './store/index'
 import { Provider } from 'react-redux'
-import {
-  createBrowserRouter,
-  RouterProvider
-} from "react-router-dom";
-import BuilderPage from './pages/builder'
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <MainPage />,
-  },
-  {
-    path: "builder",
-    element: <BuilderPage />,
-  },
-]);
-
-export default router;
+import { RouterProvider } from "react-router-dom"
+import './index.css'
+import router from './router'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <Provider store={store}>
