--- src/pages/main/style.css
@@ -1,4 +1,0 @@
-#root {
-  
-}
-
