+++ src/store/index.ts
@@ -2,9 +2,9 @@ import { configureStore } from '@reduxjs/toolkit';
 import blocksReducer from './blocks'
 
 export const store = configureStore({
-	reducer: {
-		blocks: blocksReducer
-	},
+  reducer: {
+    blocks: blocksReducer
+  },
 })
 
 export type RootState = ReturnType<typeof store.getState>
